Split constant.js file into smaller files
