+++ src/clients/artikkelit/constants/articleTypesBooks.js
@@ -1,0 +1,16 @@
+// https://wiki.eduuni.fi/display/cscsuorat/7.5+Julkaisutyypit+2021
+
+export const articleTypesBooks = [
+  {
+    value: 'A3 - Vertaisarvioitu kirjan tai muun kokoomateoksen osa',
+    text: 'A3 - Vertaisarvioitu kirjan tai muun kokoomateoksen osa'
+  },
+  {
+    value: 'B2 - (Vertaisarvioimaton) tieteellisen kirjan tai muun kokoomateoksen osa',
+    text: 'B2 - (Vertaisarvioimaton) tieteellisen kirjan tai muun kokoomateoksen osa'
+  },
+  {
+    value: 'D2 - Artikkeli ammatillisessa kokoomateoksessa',
+    text: 'D2 - Artikkeli ammatillisessa kokoomateoksessa'
+  }
+];
