+++ src/clients/artikkelit/constants/articleTypesJournal.js
@@ -1,0 +1,24 @@
+// https://wiki.eduuni.fi/display/cscsuorat/7.5+Julkaisutyypit+2021
+
+export const articleTypesJournal = [
+  {
+    value: 'A1 - Vertaisarvioitu alkuperäisartikkeli tieteellisessä aikakauslehdessä',
+    text: 'A1 - Vertaisarvioitu alkuperäisartikkeli tieteellisessä aikakauslehdessä'
+  },
+  {
+    value: 'A2 - Vertaisarvioitu katsausartikkeli tieteellisessä aikakauslehdessä',
+    text: 'A2 - Vertaisarvioitu katsausartikkeli tieteellisessä aikakauslehdessä'
+  },
+  {
+    value: 'B1 - Kirjoitus tieteellisessä aikakauslehdessä',
+    text: 'B1 - Kirjoitus tieteellisessä aikakauslehdessä'
+  },
+  {
+    value: 'D1 - Artikkeli ammattilehdessä',
+    text: 'D1 - Artikkeli ammattilehdessä'
+  },
+  {
+    value: 'E1 - Yleistajuinen artikkeli, sanomalehtiartikkeli',
+    text: 'E1 - Yleistajuinen artikkeli, sanomalehtiartikkeli'
+  }
+];
