+++ src/clients/artikkelit/constants/authorRelators.js
@@ -1,0 +1,22 @@
+export const authorRelators = [
+  {
+    value: 'kirjoittaja',
+    text: 'Kirjoittaja'
+  },
+  {
+    value: 'kuvittaja',
+    text: 'Kuvittaja'
+  },
+  {
+    value: 'kääntäjä',
+    text: 'Kääntäjä'
+  },
+  {
+    value: 'toimittaja',
+    text: 'Toimittaja'
+  },
+  {
+    value: 'yhteisö',
+    text: 'Yhteisö'
+  }
+];
