+++ src/clients/artikkelit/constants/bookTemplate.js
@@ -1,0 +1,54 @@
+export const bookTemplate = {
+  authors: [
+    {
+      asteriId: '000004321',
+      firstName: 'Malli',
+      lastName: 'Mallailija',
+      organisations: [],
+      relator: 'kirjoittaja'
+    }
+  ],
+  isElectronic: false,
+  isbns: ['123-45-6789-0'],
+  issns: [],
+  notices: [],
+  publisherInfo: {
+    publisher: 'Helsingin Mallijulkaisijat ry',
+    publisherLocation: 'Helsinki',
+    publisherYears: {
+      start: '2023'
+    }
+  },
+  publishing: 'Helsinki : Helsingin Mallijulkaisijat ry, 2023',
+  recordType: 'Monografia',
+  simpleElements: {
+    articleLanguage: ['fin'],
+    articleReference: [],
+    articleSciences: [],
+    f600s: [],
+    f610s: [],
+    f648s: [],
+    f650s: [],
+    f651s: [],
+    f653s: [],
+    f655s: [],
+    localNotes598: [],
+    methodology: [],
+    noteGeneral: [],
+    referenceLink: [],
+    titleOther: []
+  },
+  simpleGroups: {
+    abstract: {
+      language: [],
+      term: []
+    },
+    localNotes599: [],
+    otherClassificationNumber: [],
+    universalDecimalClassificationNumber: []
+  },
+  sourceIds: ['(FI-MELINDA)007654321'],
+  sourceType: 'nnam',
+  title: 'Kirjamalli',
+  years: '2023'
+};
