+++ src/clients/artikkelit/constants/ccLicenses.js
@@ -1,0 +1,34 @@
+export const ccLicenses = [
+  {
+    value: 'CC BY 4.0',
+    text: 'CC BY 4.0'
+  },
+  {
+    value: 'CC BY-NC 4.0',
+    text: 'CC BY-NC 4.0'
+  },
+  {
+    value: 'CC BY-NC-ND 4.0',
+    text: 'CC BY-NC-ND 4.0'
+  },
+  {
+    value: 'CC BY-NC-SA 4.0',
+    text: 'CC BY-NC-SA 4.0'
+  },
+  {
+    value: 'CC BY-ND 4.0',
+    text: 'CC BY-ND 4.0'
+  },
+  {
+    value: 'CC BY-SA 4.0',
+    text: 'CC BY-SA 4.0'
+  },
+  {
+    value: 'CC0 1.0',
+    text: 'CC0 1.0'
+  },
+  {
+    value: 'Public Domain 1.0',
+    text: 'Public Domain 1.0'
+  }
+]
