+++ src/clients/artikkelit/constants/index.js
@@ -1,0 +1,14 @@
+export {articleTypesBooks} from "./articleTypesBooks.js";
+export {articleTypesJournal} from "./articleTypesJournal.js";
+export {authorRelators} from "./authorRelators.js";
+export {bookTemplate} from "./bookTemplate.js";
+export {ccLicenses} from "./ccLicenses.js";
+export {journalTemplate} from "./journalTemplate.js";
+export {languages} from "./languages.js";
+export {ontologyTypes} from "./ontologyTypes.js";
+export {organizations} from "./organizations.js";
+export {sciences} from "./sciences.js";
+export {searchFilters} from "./searchFilters.js";
+export {sectionOrColumnList} from "./sectionOrColumnList.js";
+export {searchTypes} from "./searchTypes.js";
+export {sourceTypes} from "./sourceTypes.js";
