+++ src/clients/artikkelit/constants/journalTemplate.js
@@ -1,0 +1,57 @@
+export const journalTemplate = {
+  authors: [
+    {
+      asteriId: '000001234',
+      lastName: 'Malliyhteisö',
+      organisations: [],
+      relator: 'yhteisö'
+    }
+  ],
+  isElectronic: false,
+  isbns: [],
+  issns: ['0123-4567'],
+  notices: [],
+  publisherInfo: {
+    publisher: 'Mallijulkaisija',
+    publisherLocation: 'Helsinki',
+    publisherYears: {
+      start: '2022',
+      end: '2023'
+    }
+  },
+  publishing: 'Helsinki : Mallijulkaisija, 2022-2023.',
+  recordType: 'Kausijulkaisu',
+  simpleElements: {
+    articleLanguage: ['mul'],
+    articleReference: [],
+    articleSciences: [],
+    f600s: [],
+    f610s: [],
+    f648s: [],
+    f650s: [],
+    f651s: [],
+    f653s: [],
+    f655s: [],
+    localNotes598: [],
+    methodology: [],
+    noteGeneral: [],
+    referenceLink: [],
+    titleOther: [
+      'Testijulkaisu',
+      'Test publikation'
+    ]
+  },
+  simpleGroups: {
+    abstract: {
+      language: [],
+      term: []
+    },
+    localNotes599: [],
+    otherClassificationNumber: [],
+    universalDecimalClassificationNumber: []
+  },
+  sourceIds: ['(FI-MELINDA)001234567'],
+  sourceType: 'nnas',
+  title: 'Mallijulkaisu',
+  years: '2022-2023.'
+};
