+++ src/clients/artikkelit/constants/languages.js
@@ -1,0 +1,63 @@
+
+export const languages = [
+  {
+    value: 'fi;fin;Suomi',
+    text: 'Suomi'
+  },
+  {
+    value: 'en;eng;Englanti',
+    text: 'Englanti'
+  },
+  {
+    value: 'sv;swe;Ruotsi',
+    text: 'Ruotsi'
+  },
+  {
+    value: 'es;esp;Espanja',
+    text: 'Espanja'
+  },
+  {
+    value: 'it;ita;Italia',
+    text: 'Italia'
+  },
+  {
+    value: 'no;nor;Norja',
+    text: 'Norja'
+  },
+  {
+    value: 'fr;fre;Ranska',
+    text: 'Ranska'
+  },
+  {
+    value: 'de;ger;Saksa',
+    text: 'Saksa'
+  },
+  {
+    value: 'da;dan;Tanska',
+    text: 'Tanska'
+  },
+  {
+    value: 'ru;rus;Venäjä',
+    text: 'Venäjä'
+  },
+  {
+    value: 'et;est;Viro',
+    text: 'Viro'
+  },
+  {
+    value: 'smn;smn;Inarinsaame',
+    text: 'Inarinsaame'
+  },
+  {
+    value: 'sme;sme;Pohj.saame',
+    text: 'Pohjoissaame'
+  },
+  {
+    value: 'smi;smi;Muu saam.',
+    text: 'Muu saame'
+  },
+  {
+    value: 'krl;krl;Karjala',
+    text: 'Karjala'
+  }
+];
