+++ src/clients/artikkelit/constants/ontologyTypes.js
@@ -1,0 +1,75 @@
+
+export const ontologyTypes = [
+  {
+    value: 'yso',
+    text: 'YSO - Yleinen suomalainen ontologia'
+  },
+  {
+    value: 'ysoPaikat',
+    text: 'YSO-paikat'
+  },
+  {
+    value: 'ysoAika',
+    text: 'YSO-aika'
+  },
+  {
+    value: 'kanto',
+    text: 'KANTO - Kansalliset toimijatiedot'
+  },
+  {
+    value: 'slm',
+    text: 'SLM - Suomalainen lajityyppi- ja muotosanasto'
+  },
+  {
+    value: 'allfo',
+    text: 'ALLFO - Allmän finländsk ontologi '
+  },
+  {
+    value: 'allfoOrter',
+    text: 'ALLFO-orter'
+  },
+  {
+    value: 'allfoTid',
+    text: 'ALLFO-tid'
+  },
+  {
+    value: 'fgf',
+    text: 'FGF - Finländsk genre- och formlista '
+  },
+  {
+    value: 'afo',
+    text: 'AFO - Luonnonvara- ja ympäristöontologia'
+  },
+  {
+    value: 'kassu',
+    text: 'Kassu - Kasvien suomenkieliset nimet'
+  },
+  {
+    value: 'koko',
+    text: 'KOKO-ontologia'
+  },
+  {
+    value: 'finmesh',
+    text: 'MeSH / FinMeSH'
+  },
+  {
+    value: 'maotao',
+    text: 'MAO/TAO - Museoalan ja taideteollisuusalan ontologia'
+  },
+  {
+    value: 'other',
+    text: 'Muu'
+  },
+  {
+    value: 'otherPerson',
+    text: 'Muu - Henkilö'
+  },
+  {
+    value: 'otherCommunity',
+    text: 'Muu - Yhteisö'
+  },
+  {
+    value: 'otherTime',
+    text: 'Muu - Aikamääre'
+  }
+];
