+++ src/clients/artikkelit/constants/organizations.js
@@ -1,0 +1,61 @@
+// https://wiki.eduuni.fi/display/cscsuorat/7.6+Tutkimuslaitosten+ja+yliopistollisten+sairaaloiden+organisaatiotunnukset+2022
+// https://wiki.eduuni.fi/display/cscsuorat/7.1+Korkeakoulujen+oppilaitosnumerot+2022
+// Latest update: 24.8.2023 (MUU-384)
+
+export const organizations = [
+  {value: 'Aalto yliopisto', text: '', code: '10076'},
+  {value: 'Centria-ammattikorkeakoulu', text: '', code: '02536'},
+  {value: 'Diakonia-ammattikorkeakoulu', text: '', code: '02623'},
+  {value: 'Geologian tutkimuskeskus', text: '', code: '5040011'},
+  {value: 'Haaga-Helia ammattikorkeakoulu', text: '', code: '10056'},
+  {value: 'Helsingin seudun yliopistollisen keskussairaalan erityisvastuualue', text: '', code: '15675350'},
+  {value: 'Helsingin yliopisto', text: '', code: '01901'},
+  {value: 'Humanistinen ammattikorkeakoulu', text: '', code: '02631'},
+  {value: 'Hämeen ammattikorkeakoulu', text: '', code: '02467'},
+  {value: 'Ilmatieteen laitos', text: '', code: '4940015'},
+  {value: 'Itä-Suomen yliopisto', text: '', code: '10088'},
+  {value: 'Jyväskylän ammattikorkeakoulu', text: '', code: '02504'},
+  {value: 'Jyväskylän yliopisto', text: '', code: '01906'},
+  {value: 'Kaakkois-Suomen ammattikorkeakoulu', text: '', code: '10118'},
+  {value: 'Kajaanin ammattikorkeakoulu', text: '', code: '02473'},
+  {value: 'Karelia-ammattikorkeakoulu', text: '', code: '02469'},
+  {value: 'Kuopion yliopistollisen sairaalan erityisvastuualue', text: '', code: '01714953'},
+  {value: 'LAB-ammattikorkeakoulu', text: '', code: '10126'},
+  {value: 'Lapin ammattikorkeakoulu', text: '', code: '10108'},
+  {value: 'Lapin yliopisto', text: '', code: '01918'},
+  {value: 'Lappeenrannan-Lahden teknillinen yliopisto LUT', text: '', code: '01914'},
+  {value: 'Laurea-ammattikorkeakoulu', text: '', code: '02629'},
+  {value: 'Luonnonvarakeskus', text: '', code: '4100010'},
+  {value: 'Lääkealan turvallisuus- ja kehittämiskeskus', text: '', code: '558005'},
+  {value: 'Maanmittauslaitos', text: '', code: '4020217'},
+  {value: 'Metropolia ammattikorkeakoulu', text: '', code: '10065'},
+  {value: 'Oulun ammattikorkeakoulu', text: '', code: '02471'},
+  {value: 'Oulun yliopisto', text: '', code: '01904'},
+  {value: 'Oulun yliopistollisen sairaalan erityisvastuualue', text: '', code: '06794809'},
+  {value: 'Poliisiammattikorkeakoulu', text: '', code: '02557'},
+  {value: 'Ruokavirasto', text: '', code: '430001'},
+  {value: 'Satakunnan ammattikorkeakoulu', text: '', code: '02507'},
+  {value: 'Savonia-ammattikorkeakoulu', text: '', code: '02537'},
+  {value: 'Seinäjoen ammattikorkeakoulu', text: '', code: '02472'},
+  {value: 'Suomen Pankki', text: '', code: '02022481'},
+  {value: 'Suomen ympäristökeskus', text: '', code: '7020017'},
+  {value: 'Svenska handelshögskolan', text: '', code: '01910'},
+  {value: 'Säteilyturvakeskus', text: '', code: '5550012'},
+  {value: 'Taideyliopisto', text: '', code: '10103'},
+  {value: 'Tampereen ammattikorkeakoulu', text: '', code: '02630'},
+  {value: 'Tampereen yliopisto', text: '', code: '10122'},
+  {value: 'Tampereen yliopistollisen sairaalan erityisvastuualue', text: '', code: '08265978'},
+  {value: 'Teknologian tutkimuskeskus VTT Oy', text: '', code: '26473754'},
+  {value: 'Terveyden ja hyvinvoinnin laitos', text: '', code: '5610017'},
+  {value: 'Turun ammattikorkeakoulu', text: '', code: '02509'},
+  {value: 'Turun yliopisto', text: '', code: '10089'},
+  {value: 'Turun yliopistollisen keskussairaalan erityisvastuualue', text: '', code: '08282559'},
+  {value: 'Työterveyslaitos', text: '', code: '02202669'},
+  {value: 'Ulkopoliittinen instituutti', text: '', code: '1120017'},
+  {value: 'Vaasan ammattikorkeakoulu', text: '', code: '02627'},
+  {value: 'Vaasan yliopisto  ', text: '', code: '01913'},
+  {value: 'Valtion taloudellinen tutkimuskeskus', text: '', code: '3060016'},
+  {value: 'Yrkeshögskolan Arcada', text: '', code: '02535'},
+  {value: 'Yrkeshögskolan Novia', text: '', code: '10066'},
+  {value: 'Åbo Akademi', text: '', code: '01903'},
+];
