+++ src/clients/artikkelit/constants/sciences.js
@@ -1,0 +1,268 @@
+// https://wiki.eduuni.fi/display/cscsuorat/7.3+Tieteenalaluokitus+2021
+
+export const sciences = [
+  {
+    value: '1 - Luonnontieteet - 111 - Matematiikka',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 112 - Tilastotiede',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 113 - Tietojenkäsittely ja informaatiotieteet',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 114 - Fysiikka',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 115 - Avaruustieteet ja tähtitiede',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 116 - Kemia',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 1171 - Geotieteet',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 1172 - Ympäristötiede',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 1181 - Ekologia, evoluutiobiologia',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 1182 - Biokemia, solu- ja molekyylibiologia',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 1183 - Kasvibiologia, mikrobiologia, virologia',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 1184 - Genetiikka, kehitysbiologia, fysiologia',
+    text: ''
+  },
+  {
+    value: '1 - Luonnontieteet - 119 - Muut luonnontieteet',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 211 - Arkkitehtuuri',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 212 - Rakennus- ja yhdyskuntatekniikka',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 213 - Sähkö-, automaatio- ja tietoliikennetekniikka, elektroniikka',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 214 - Kone- ja valmistustekniikka',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 215 - Teknillinen kemia, kemian prosessitekniikka',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 216 - Materiaalitekniikka',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 217 - Lääketieteen tekniikka',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 218 - Ympäristötekniikka',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 219 - Ympäristön bioteknologia',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 220 - Teollinen bioteknologia',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 221 - Nanoteknologia',
+    text: ''
+  },
+  {
+    value: '2 - Tekniikka - 222 - Muu tekniikka',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 3111 - Biolääketieteet',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 3112 - Neurotieteet',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 3121 - Yleislääketiede, sisätaudit ja muut kliiniset lääketieteet',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 3122 - Syöpätaudit',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 3123 - Naisten- ja lastentaudit',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 3124 - Neurologia ja psykiatria',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 3125 - Korva-, nenä- ja kurkkutaudit, silmätaudit',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 3126 - Kirurgia, anestesiologia, tehohoito, radiologia',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 313 - Hammaslääketieteet',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 3141 - Terveystiede',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 3142 - Kansanterveystiede, ympäristö ja työterveys',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 315 - Liikuntatiede',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 316 - Hoitotiede',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 317 - Farmasia',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 318 - Lääketieteen bioteknologia',
+    text: ''
+  },
+  {
+    value: '3 - Lääke- ja terveystieteet - 319 - Oikeuslääketiede ja muut lääketieteet',
+    text: ''
+  },
+  {
+    value: '4 - Maatalous- ja metsätieteet - 4111 - Maataloustiede',
+    text: ''
+  },
+  {
+    value: '4 - Maatalous- ja metsätieteet - 4112 - Metsätiede',
+    text: ''
+  },
+  {
+    value: '4 - Maatalous- ja metsätieteet - 412 - Kotieläintiede, maitotaloustiede',
+    text: ''
+  },
+  {
+    value: '4 - Maatalous- ja metsätieteet - 413 - Eläinlääketiede',
+    text: ''
+  },
+  {
+    value: '4 - Maatalous- ja metsätieteet - 414 - Maatalouden bioteknologia',
+    text: ''
+  },
+  {
+    value: '4 - Maatalous- ja metsätieteet - 415 - Muut maataloustieteet',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 511 - Kansantaloustiede',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 512 - Liiketaloustiede',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 513 - Oikeustiede',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 5141 - Sosiologia',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 5142 - Sosiaali- ja yhteiskuntapolitiikka',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 515 - Psykologia',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 516 - Kasvatustieteet',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 517 - Valtio-oppi, hallintotiede',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 518 - Media- ja viestintätieteet',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 519 - Yhteiskuntamaantiede, talousmaantiede',
+    text: ''
+  },
+  {
+    value: '5 - Yhteiskuntatieteet - 520 - Muut yhteiskuntatieteet',
+    text: ''
+  },
+  {
+    value: '6 - Humanistiset tieteet - 611 - Filosofia',
+    text: ''
+  },
+  {
+    value: '6 - Humanistiset tieteet - 6121 - Kielitieteet',
+    text: ''
+  },
+  {
+    value: '6 - Humanistiset tieteet - 6122 - Kirjallisuuden tutkimus',
+    text: ''
+  },
+  {
+    value: '6 - Humanistiset tieteet - 6131 - Teatteri, tanssi, musiikki, muut esittävät taiteet',
+    text: ''
+  },
+  {
+    value: '6 - Humanistiset tieteet - 6132 - Kuvataide ja muotoilu',
+    text: ''
+  },
+  {
+    value: '6 - Humanistiset tieteet - 614 - Teologia',
+    text: ''
+  },
+  {
+    value: '6 - Humanistiset tieteet - 615 - Historia ja arkeologia',
+    text: ''
+  },
+  {
+    value: '6 - Humanistiset tieteet - 616 - Muut humanistiset tieteet',
+    text: ''
+  }
+];
