+++ src/clients/artikkelit/constants/searchFilters.js
@@ -1,0 +1,14 @@
+export const searchFilters = [
+  {
+    value: 'filter-arto',
+    text: 'Arto-tietueet'
+  },
+  {
+    value: 'filter-arto-fennica',
+    text: 'Arto- ja Fennica-tietueet'
+  },
+  {
+    value: 'filter-arto-fennica-melinda',
+    text: 'Arto-, Fennica- ja Melinda-tietueet'
+  }
+];
