+++ src/clients/artikkelit/constants/searchTypes.js
@@ -1,0 +1,18 @@
+export const searchTypes = [
+  {
+    value: 'title',
+    text: 'Nimeke'
+  },
+  {
+    value: 'melinda',
+    text: 'Melinda-ID'
+  },
+  {
+    value: 'issn',
+    text: 'ISSN'
+  },
+  {
+    value: 'isbn',
+    text: 'ISBN'
+  }
+];
