+++ src/clients/artikkelit/constants/sectionOrColumnList.js
@@ -1,0 +1,18 @@
+export const sectionOrColumnList = [
+  {
+    value: 'Artikkelit',
+    text: ''
+  },
+  {
+    value: 'Kirjallisuus',
+    text: ''
+  },
+  {
+    value: 'Muut',
+    text: ''
+  },
+  {
+    value: 'Vertaisarvioidut artikkelit',
+    text: ''
+  }
+];
