+++ src/clients/artikkelit/constants/sourceTypes.js
@@ -1,0 +1,10 @@
+export const sourceTypes = [
+  {
+    value: 'journal',
+    text: 'Lehtiartikkeli'
+  },
+  {
+    value: 'book',
+    text: 'Artikkeli kokoomateoksessa'
+  }
+];
